pyproject.toml
[project]
name = "nexus"
version = "2.0.0"
description = "Backend API for Z3R0"
authors = [
    {name = "null2264", email = "palembani@gmail.com"}
]
license = {text = "MPL-2.0"}
requires-python = ">=3.10"
dependencies = [
    "fastapi==0.100.0",
    "discord.py>=2.2",
    "pyzmq==25.0.0",
    "websockets>=10.4",
    "uvicorn>=0.20.0",
    "python-dotenv==1.0.0",
    "itsdangerous>=2.1.2",
    "oauthlib>=3.2.2",
    "beautifulsoup4>=4.12.2",
    "Pillow>=10.0.0",
    "pydantic>=1.10,<=2.0",
    "redis>=4.6.0",
]

[tool.pdm.scripts]
start = "uvicorn nexus.app:app --reload"

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.black]
line-length = 125
skip-string-normalization = true

[tool.isort]
profile = "black"
multi_line_output = 3
lines_after_imports = 2

[tool.pyright]
reportUnnecessaryTypeIgnoreComment = "warning"
reportUnusedImport = "error"
pythonVersion = "3.9"
typeCheckingMode = "basic"

[tool.pylint.main]
extension-pkg-allow-list = ["pydantic"]
